Prepare UserWalletRepository, UserLoginRepository, and refactor.
Refactor:
- change wording in log message
- re-structure package
- organize imports

Add new thing:
- prepare prepare UserWalletRepository, and UserLoginRepository
- create `ResponseAdviceHandler.handleCredentialExistsException`
