+++ src/main/java/dev/fResult/goutTogether/auths/UserLoginRepository.java
@@ -1,5 +1,9 @@
 package dev.fResult.goutTogether.auths;
 
-import org.springframework.data.repository.CrudRepository;
+import dev.fResult.goutTogether.auths.entities.UserLogin;
+import java.util.Optional;
+import org.springframework.data.repository.ListCrudRepository;
 
-public interface UserLoginRepository extends CrudRepository<UserLogin, Integer> {}
+public interface UserLoginRepository extends ListCrudRepository<UserLogin, Integer> {
+    Optional<UserLogin> findOneByEmail(String email);
+}
