+++ src/main/java/dev/fResult/goutTogether/auths/services/AuthService.java
@@ -1,6 +1,6 @@
 package dev.fResult.goutTogether.auths.services;
 
-import dev.fResult.goutTogether.auths.UserLogin;
+import dev.fResult.goutTogether.auths.entities.UserLogin;
 
 public interface AuthService {
   UserLogin findUserCredentialByEmail(String email);
