+++ src/main/java/dev/fResult/goutTogether/tourCompanies/repositories/TourCompanyLoginRepository.java
@@ -1,6 +1,6 @@
 package dev.fResult.goutTogether.tourCompanies.repositories;
 
-import dev.fResult.goutTogether.tourCompanies.entities.TourCompanyLogin;
+import dev.fResult.goutTogether.auths.entities.TourCompanyLogin;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
