+++ src/main/java/dev/fResult/goutTogether/wallets/entities/TourCompanyWallet.java
@@ -1,12 +1,11 @@
 package dev.fResult.goutTogether.wallets.entities;
 
 import dev.fResult.goutTogether.tourCompanies.entities.TourCompany;
+import java.math.BigDecimal;
+import java.time.Instant;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Table;
-
-import java.math.BigDecimal;
-import java.time.Instant;
 
 @Table("tour_company_wallets")
 public record TourCompanyWallet(
