+++ src/main/java/dev/fResult/goutTogether/wallets/repositories/UserWalletRepository.java
@@ -1,4 +1,6 @@
 package dev.fResult.goutTogether.wallets.repositories;
 
-public interface UserWalletRepository {
-}
+import dev.fResult.goutTogether.wallets.entities.UserWallet;
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserWalletRepository extends CrudRepository<UserWallet, Integer> {}
